+++ config/config.go
@@ -13,7 +13,6 @@ import (
 	"runtime"
 	"strconv"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/thrasher-corp/gocryptotrader/common"
@@ -25,67 +24,6 @@ import (
 	"github.com/thrasher-corp/gocryptotrader/database"
 	"github.com/thrasher-corp/gocryptotrader/exchanges/asset"
 	log "github.com/thrasher-corp/gocryptotrader/logger"
-)
-
-// Constants declared here are filename strings and test strings
-const (
-	FXProviderFixer                            = "fixer"
-	EncryptedConfigFile                        = "config.dat"
-	ConfigFile                                 = "config.json"
-	ConfigTestFile                             = "../testdata/configtest.json"
-	configFileEncryptionPrompt                 = 0
-	configFileEncryptionEnabled                = 1
-	configFileEncryptionDisabled               = -1
-	configPairsLastUpdatedWarningThreshold     = 30 // 30 days
-	configDefaultHTTPTimeout                   = time.Second * 15
-	configDefaultWebsocketResponseCheckTimeout = time.Millisecond * 30
-	configDefaultWebsocketResponseMaxLimit     = time.Second * 7
-	configDefaultWebsocketOrderbookBufferLimit = 5
-	configDefaultWebsocketTrafficTimeout       = time.Second * 30
-	configMaxAuthFailures                      = 3
-	defaultNTPAllowedDifference                = 50000000
-	defaultNTPAllowedNegativeDifference        = 50000000
-
-	DefaultAPIKey      = "Key"
-	DefaultAPISecret   = "Secret"
-	DefaultAPIClientID = "ClientID"
-)
-
-// Constants here hold some messages
-const (
-	ErrExchangeNameEmpty                       = "exchange #%d name is empty"
-	ErrExchangeAvailablePairsEmpty             = "exchange %s available pairs is empty"
-	ErrExchangeEnabledPairsEmpty               = "exchange %s enabled pairs is empty"
-	ErrExchangeBaseCurrenciesEmpty             = "exchange %s base currencies is empty"
-	ErrExchangeNotFound                        = "exchange %s not found"
-	ErrNoEnabledExchanges                      = "no exchanges enabled"
-	ErrCryptocurrenciesEmpty                   = "cryptocurrencies variable is empty"
-	ErrFailureOpeningConfig                    = "fatal error opening %s file. Error: %s"
-	ErrCheckingConfigValues                    = "fatal error checking config values. Error: %s"
-	ErrSavingConfigBytesMismatch               = "config file %q bytes comparison doesn't match, read %s expected %s"
-	WarningWebserverCredentialValuesEmpty      = "webserver support disabled due to empty Username/Password values"
-	WarningWebserverListenAddressInvalid       = "webserver support disabled due to invalid listen address"
-	WarningExchangeAuthAPIDefaultOrEmptyValues = "exchange %s authenticated API support disabled due to default/empty APIKey/Secret/ClientID values"
-	WarningPairsLastUpdatedThresholdExceeded   = "exchange %s last manual update of available currency pairs has exceeded %d days. Manual update required!"
-)
-
-// Constants here define unset default values displayed in the config.json
-// file
-const (
-	APIURLNonDefaultMessage              = "NON_DEFAULT_HTTP_LINK_TO_EXCHANGE_API"
-	WebsocketURLNonDefaultMessage        = "NON_DEFAULT_HTTP_LINK_TO_WEBSOCKET_EXCHANGE_API"
-	DefaultUnsetAPIKey                   = "Key"
-	DefaultUnsetAPISecret                = "Secret"
-	DefaultUnsetAccountPlan              = "accountPlan"
-	DefaultForexProviderExchangeRatesAPI = "ExchangeRates"
-)
-
-// Variables here are used for configuration
-var (
-	Cfg            Config
-	IsInitialSetup bool
-	testBypass     bool
-	m              sync.Mutex
 )
 
 // GetCurrencyConfig returns currency configurations
@@ -1066,9 +1004,6 @@ func (c *Config) CheckExchangeConfigValues() error {
 // CheckCurrencyConfigValues checks to see if the currency config values are correct or not
 func (c *Config) CheckCurrencyConfigValues() error {
 	fxProviders := forexprovider.GetAvailableForexProviders()
-	if len(fxProviders) == 0 {
-		return errors.New("no forex providers available")
-	}
 
 	if len(fxProviders) != len(c.Currency.ForexProviders) {
 		for x := range fxProviders {
@@ -1087,34 +1022,32 @@ func (c *Config) CheckCurrencyConfigValues() error {
 
 	count := 0
 	for i := range c.Currency.ForexProviders {
-		if c.Currency.ForexProviders[i].Enabled {
-			if c.Currency.ForexProviders[i].APIKey == DefaultUnsetAPIKey && c.Currency.ForexProviders[i].Name != DefaultForexProviderExchangeRatesAPI {
-				log.Warnf(log.Global, "%s enabled forex provider API key not set. Please set this in your config.json file\n", c.Currency.ForexProviders[i].Name)
-				c.Currency.ForexProviders[i].Enabled = false
-				c.Currency.ForexProviders[i].PrimaryProvider = false
-				continue
-			}
-
-			if c.Currency.ForexProviders[i].Name == "CurrencyConverter" {
-				if c.Currency.ForexProviders[i].Enabled &&
-					c.Currency.ForexProviders[i].PrimaryProvider &&
-					(c.Currency.ForexProviders[i].APIKey == "" ||
-						c.Currency.ForexProviders[i].APIKey == DefaultUnsetAPIKey) {
-					log.Warnln(log.Global, "CurrencyConverter forex provider no longer supports unset API key requests. Switching to ExchangeRates FX provider..")
-					c.Currency.ForexProviders[i].Enabled = false
-					c.Currency.ForexProviders[i].PrimaryProvider = false
-					c.Currency.ForexProviders[i].APIKey = DefaultUnsetAPIKey
-					c.Currency.ForexProviders[i].APIKeyLvl = -1
-					continue
-				}
-			}
-
-			if c.Currency.ForexProviders[i].APIKeyLvl == -1 && c.Currency.ForexProviders[i].Name != DefaultForexProviderExchangeRatesAPI {
-				log.Warnf(log.Global, "%s APIKey Level not set, functions limited. Please set this in your config.json file\n",
-					c.Currency.ForexProviders[i].Name)
-			}
-			count++
+		if c.Currency.ForexProviders[i].Name == "CurrencyConverter" &&
+			c.Currency.ForexProviders[i].Enabled &&
+			c.Currency.ForexProviders[i].PrimaryProvider &&
+			(c.Currency.ForexProviders[i].APIKey == "" ||
+				c.Currency.ForexProviders[i].APIKey == DefaultUnsetAPIKey) {
+			log.Warnln(log.Global, "CurrencyConverter forex provider no longer supports unset API key requests. Switching to ExchangeRates FX provider..")
+			c.Currency.ForexProviders[i].Enabled = false
+			c.Currency.ForexProviders[i].PrimaryProvider = false
+			c.Currency.ForexProviders[i].APIKey = DefaultUnsetAPIKey
+			c.Currency.ForexProviders[i].APIKeyLvl = -1
+			continue
 		}
+		if c.Currency.ForexProviders[i].Enabled &&
+			c.Currency.ForexProviders[i].APIKey == DefaultUnsetAPIKey &&
+			c.Currency.ForexProviders[i].Name != DefaultForexProviderExchangeRatesAPI {
+			log.Warnf(log.Global, "%s enabled forex provider API key not set. Please set this in your config.json file\n", c.Currency.ForexProviders[i].Name)
+			c.Currency.ForexProviders[i].Enabled = false
+			c.Currency.ForexProviders[i].PrimaryProvider = false
+			continue
+		}
+
+		if c.Currency.ForexProviders[i].APIKeyLvl == -1 && c.Currency.ForexProviders[i].Name != DefaultForexProviderExchangeRatesAPI {
+			log.Warnf(log.Global, "%s APIKey Level not set, functions limited. Please set this in your config.json file\n",
+				c.Currency.ForexProviders[i].Name)
+		}
+		count++
 	}
 
 	if count == 0 {
