+++ config/config_types.go
@@ -3,6 +3,7 @@ package config
 import (
 	"fmt"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/thrasher-corp/gocryptotrader/currency"
@@ -10,6 +11,67 @@ import (
 	"github.com/thrasher-corp/gocryptotrader/database"
 	log "github.com/thrasher-corp/gocryptotrader/logger"
 	"github.com/thrasher-corp/gocryptotrader/portfolio"
+)
+
+// Constants declared here are filename strings and test strings
+const (
+	FXProviderFixer                            = "fixer"
+	EncryptedConfigFile                        = "config.dat"
+	ConfigFile                                 = "config.json"
+	ConfigTestFile                             = "../testdata/configtest.json"
+	configFileEncryptionPrompt                 = 0
+	configFileEncryptionEnabled                = 1
+	configFileEncryptionDisabled               = -1
+	configPairsLastUpdatedWarningThreshold     = 30 // 30 days
+	configDefaultHTTPTimeout                   = time.Second * 15
+	configDefaultWebsocketResponseCheckTimeout = time.Millisecond * 30
+	configDefaultWebsocketResponseMaxLimit     = time.Second * 7
+	configDefaultWebsocketOrderbookBufferLimit = 5
+	configDefaultWebsocketTrafficTimeout       = time.Second * 30
+	configMaxAuthFailures                      = 3
+	defaultNTPAllowedDifference                = 50000000
+	defaultNTPAllowedNegativeDifference        = 50000000
+
+	DefaultAPIKey      = "Key"
+	DefaultAPISecret   = "Secret"
+	DefaultAPIClientID = "ClientID"
+)
+
+// Constants here hold some messages
+const (
+	ErrExchangeNameEmpty                       = "exchange #%d name is empty"
+	ErrExchangeAvailablePairsEmpty             = "exchange %s available pairs is empty"
+	ErrExchangeEnabledPairsEmpty               = "exchange %s enabled pairs is empty"
+	ErrExchangeBaseCurrenciesEmpty             = "exchange %s base currencies is empty"
+	ErrExchangeNotFound                        = "exchange %s not found"
+	ErrNoEnabledExchanges                      = "no exchanges enabled"
+	ErrCryptocurrenciesEmpty                   = "cryptocurrencies variable is empty"
+	ErrFailureOpeningConfig                    = "fatal error opening %s file. Error: %s"
+	ErrCheckingConfigValues                    = "fatal error checking config values. Error: %s"
+	ErrSavingConfigBytesMismatch               = "config file %q bytes comparison doesn't match, read %s expected %s"
+	WarningWebserverCredentialValuesEmpty      = "webserver support disabled due to empty Username/Password values"
+	WarningWebserverListenAddressInvalid       = "webserver support disabled due to invalid listen address"
+	WarningExchangeAuthAPIDefaultOrEmptyValues = "exchange %s authenticated API support disabled due to default/empty APIKey/Secret/ClientID values"
+	WarningPairsLastUpdatedThresholdExceeded   = "exchange %s last manual update of available currency pairs has exceeded %d days. Manual update required!"
+)
+
+// Constants here define unset default values displayed in the config.json
+// file
+const (
+	APIURLNonDefaultMessage              = "NON_DEFAULT_HTTP_LINK_TO_EXCHANGE_API"
+	WebsocketURLNonDefaultMessage        = "NON_DEFAULT_HTTP_LINK_TO_WEBSOCKET_EXCHANGE_API"
+	DefaultUnsetAPIKey                   = "Key"
+	DefaultUnsetAPISecret                = "Secret"
+	DefaultUnsetAccountPlan              = "accountPlan"
+	DefaultForexProviderExchangeRatesAPI = "ExchangeRates"
+)
+
+// Variables here are used for configuration
+var (
+	Cfg            Config
+	IsInitialSetup bool
+	testBypass     bool
+	m              sync.Mutex
 )
 
 // Config is the overarching object that holds all the information for
