+++ currency/forexprovider/base/base.go
@@ -4,6 +4,21 @@ import (
 	"time"
 )
 
+// GetName returns name of provider
+func (b *Base) GetName() string {
+	return b.Name
+}
+
+// IsEnabled returns true if enabled
+func (b *Base) IsEnabled() bool {
+	return b.Enabled
+}
+
+// IsPrimaryProvider returns true if primary provider
+func (b *Base) IsPrimaryProvider() bool {
+	return b.PrimaryProvider
+}
+
 // DefaultTimeOut is the default timeout for foreign exchange providers
 const DefaultTimeOut = time.Second * 15
 
@@ -21,19 +36,4 @@ type Settings struct {
 // Base enforces standard variables across the provider packages
 type Base struct {
 	Settings `json:"settings"`
-}
-
-// GetName returns name of provider
-func (b *Base) GetName() string {
-	return b.Name
-}
-
-// IsEnabled returns true if enabled
-func (b *Base) IsEnabled() bool {
-	return b.Enabled
-}
-
-// IsPrimaryProvider returns true if primary provider
-func (b *Base) IsPrimaryProvider() bool {
-	return b.PrimaryProvider
 }
