+++ currency/forexprovider/currencyconverterapi/currencyconverterapi_types.go
@@ -1,5 +1,33 @@
 package currencyconverter
 
+import (
+	"github.com/thrasher-corp/gocryptotrader/currency/forexprovider/base"
+	"github.com/thrasher-corp/gocryptotrader/exchanges/request"
+)
+
+// const declarations consist of endpoints
+const (
+	APIEndpointURL     = "https://currencyconverterapi.com/api/"
+	APIEndpointFreeURL = "https://free.currencyconverterapi.com/api/"
+	APIEndpointVersion = "v5"
+
+	APIEndpointConvert    = "convert"
+	APIEndpointCurrencies = "currencies"
+	APIEndpointCountries  = "countries"
+	APIEndpointUsage      = "usage"
+
+	defaultAPIKey = "Key"
+
+	authRate   = 0
+	unAuthRate = 0
+)
+
+// CurrencyConverter stores the struct for the CurrencyConverter API
+type CurrencyConverter struct {
+	base.Base
+	Requester *request.Requester
+}
+
 // Error stores the error message
 type Error struct {
 	Status int    `json:"status"`
