+++ currency/forexprovider/currencylayer/currencylayer.go
@@ -26,34 +26,6 @@ import (
 	log "github.com/thrasher-corp/gocryptotrader/logger"
 )
 
-// const declarations consist of endpoints and APIKey privileges
-const (
-	AccountFree = iota
-	AccountBasic
-	AccountPro
-	AccountEnterprise
-
-	APIEndpointURL        = "http://apilayer.net/api/"
-	APIEndpointURLSSL     = "https://apilayer.net/api/"
-	APIEndpointList       = "list"
-	APIEndpointLive       = "live"
-	APIEndpointHistorical = "historical"
-	APIEndpointConversion = "convert"
-	APIEndpointTimeframe  = "timeframe"
-	APIEndpointChange     = "change"
-
-	authRate   = 0
-	unAuthRate = 0
-)
-
-// CurrencyLayer is a foreign exchange rate provider at
-// https://currencylayer.com NOTE default base currency is USD when using a free
-// account. Has automatic upgrade to a SSL connection.
-type CurrencyLayer struct {
-	base.Base
-	Requester *request.Requester
-}
-
 // Setup sets appropriate values for CurrencyLayer
 func (c *CurrencyLayer) Setup(config base.Settings) error {
 	if config.APIKeyLvl < 0 || config.APIKeyLvl > 3 {
