+++ currency/forexprovider/exchangeratesapi.io/exchangeratesapi_types.go
@@ -1,5 +1,28 @@
 package exchangerates
 
+import (
+	"github.com/thrasher-corp/gocryptotrader/currency/forexprovider/base"
+	"github.com/thrasher-corp/gocryptotrader/exchanges/request"
+)
+
+const (
+	exchangeRatesAPI                 = "https://api.exchangeratesapi.io"
+	exchangeRatesLatest              = "latest"
+	exchangeRatesHistory             = "history"
+	exchangeRatesSupportedCurrencies = "EUR,CHF,USD,BRL,ISK,PHP,KRW,BGN,MXN," +
+		"RON,CAD,SGD,NZD,THB,HKD,JPY,NOK,HRK,ILS,GBP,DKK,HUF,MYR,RUB,TRY,IDR," +
+		"ZAR,INR,AUD,CZK,SEK,CNY,PLN"
+
+	authRate   = 0
+	unAuthRate = 0
+)
+
+// ExchangeRates stores the struct for the ExchangeRatesAPI API
+type ExchangeRates struct {
+	base.Base
+	Requester *request.Requester
+}
+
 // Rates holds the latest forex rates info
 type Rates struct {
 	Base  string             `json:"base"`
