+++ currency/forexprovider/forexprovider.go
@@ -13,11 +13,6 @@ import (
 	"github.com/thrasher-corp/gocryptotrader/currency/forexprovider/openexchangerates"
 )
 
-// ForexProviders is a foreign exchange handler type
-type ForexProviders struct {
-	base.FXHandler
-}
-
 // GetAvailableForexProviders returns a list of supported forex providers
 func GetAvailableForexProviders() []string {
 	return []string{"CurrencyConverter",
@@ -134,4 +129,9 @@ func StartFXService(fxProviders []base.Settings) (*ForexProviders, error) {
 	}
 
 	return handler, nil
+}
+
+// ForexProviders is a foreign exchange handler type
+type ForexProviders struct {
+	base.FXHandler
 }
