+++ currency/forexprovider/openexchangerates/openexchangerates.go
@@ -23,45 +23,6 @@ import (
 	log "github.com/thrasher-corp/gocryptotrader/logger"
 )
 
-// These consts contain endpoint information
-const (
-	APIDeveloperAccess = iota
-	APIEnterpriseAccess
-	APIUnlimitedAccess
-
-	APIURL                = "https://openexchangerates.org/api/"
-	APIEndpointLatest     = "latest.json"
-	APIEndpointHistorical = "historical/%s.json"
-	APIEndpointCurrencies = "currencies.json"
-	APIEndpointTimeSeries = "time-series.json"
-	APIEndpointConvert    = "convert/%s/%s/%s"
-	APIEndpointOHLC       = "ohlc.json"
-	APIEndpointUsage      = "usage.json"
-
-	oxrSupportedCurrencies = "AED,AFN,ALL,AMD,ANG,AOA,ARS,AUD,AWG,AZN,BAM,BBD," +
-		"BDT,BGN,BHD,BIF,BMD,BND,BOB,BRL,BSD,BTC,BTN,BWP,BYN,BYR,BZD,CAD,CDF," +
-		"CHF,CLF,CLP,CNH,CNY,COP,CRC,CUC,CUP,CVE,CZK,DJF,DKK,DOP,DZD,EEK,EGP," +
-		"ERN,ETB,EUR,FJD,FKP,GBP,GEL,GGP,GHS,GIP,GMD,GNF,GTQ,GYD,HKD,HNL,HRK," +
-		"HTG,HUF,IDR,ILS,IMP,INR,IQD,IRR,ISK,JEP,JMD,JOD,JPY,KES,KGS,KHR,KMF," +
-		"KPW,KRW,KWD,KYD,KZT,LAK,LBP,LKR,LRD,LSL,LYD,MAD,MDL,MGA,MKD,MMK,MNT," +
-		"MOP,MRO,MRU,MTL,MUR,MVR,MWK,MXN,MYR,MZN,NAD,NGN,NIO,NOK,NPR,NZD,OMR," +
-		"PAB,PEN,PGK,PHP,PKR,PLN,PYG,QAR,RON,RSD,RUB,RWF,SAR,SBD,SCR,SDG,SEK," +
-		"SGD,SHP,SLL,SOS,SRD,SSP,STD,STN,SVC,SYP,SZL,THB,TJS,TMT,TND,TOP,TRY," +
-		"TTD,TWD,TZS,UAH,UGX,USD,UYU,UZS,VEF,VND,VUV,WST,XAF,XAG,XAU,XCD,XDR," +
-		"XOF,XPD,XPF,XPT,YER,ZAR,ZMK,ZMW"
-
-	authRate   = 0
-	unAuthRate = 0
-)
-
-// OXR is a foreign exchange rate provider at https://openexchangerates.org/
-// this is the overarching type across this package
-// DOCs : https://docs.openexchangerates.org/docs
-type OXR struct {
-	base.Base
-	Requester *request.Requester
-}
-
 // Setup sets values for the OXR object
 func (o *OXR) Setup(config base.Settings) error {
 	if config.APIKeyLvl < 0 || config.APIKeyLvl > 2 {
